1. Enable Network stack to pass SCT, currently MNP, ARP, IP4, TCP4 and DHCP4 have passed SCT. 2. Enable Network stack to pass ICC build. 3. Sync Network library instances' module names with their filenames.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@3692 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Universal/Network/Ip4Dxe/Ip4Igmp.c
@@ -206,7 +206,7 @@ Ip4SendIgmpMessage (
   Igmp->MaxRespTime = 0;
   Igmp->Checksum    = 0;
   Igmp->Group       = HTONL (Group);
-  Igmp->Checksum    = ~NetblockChecksum ((UINT8 *) Igmp, sizeof (IGMP_HEAD));
+  Igmp->Checksum    = (UINT16) (~NetblockChecksum ((UINT8 *) Igmp, sizeof (IGMP_HEAD)));
 
   Head.Tos          = 0;
   Head.Protocol     = IP4_PROTO_IGMP;
